Split CLI into library and binary part
Enables other executables to invoke dora CLI commands directly without going through the argument parsing.
